Update containerd/containerd indirect test dep
This dependency is pulled in from our testing infra and not in our final
Vault version. However, updating this dep pulls in newer versions of
other deps (such as protobuf) which are used at runtime. Updated via:

$ go get -u github.com/containerd/containerd@v1.5.12 && go mod tidy

Since docker/docker has an indirect dep on containerd, I've updated it
as well:

$ go get -u github.com/docker/docker@v20.10.16 && go mod tidy

Signed-off-by: Alexander Scheel <alex.scheel@hashicorp.com>